feat: #68 - CoreML 관련 레포지토리 구현
- CoreML 데이터 추출 후 Entity 배열로 변환하는 기능 구현
- Domain에서 interface 구현
- Data Layer에서 필요한 내용은 DataLayer에서 구현
