+++ BestWish/Data/Util/UIImage+Util.swift
@@ -1,0 +1,50 @@
+//
+//  UIImage+UIImage.swift
+//  BestWish
+//
+//  Created by Quarang on 6/17/25.
+//
+
+import UIKit
+
+// MARK: - CoreML에서 이미지 형식 변화
+extension UIImage {
+    /// CoreML 이미지 입력 형식 변환
+    func toCVPixelBuffer() -> CVPixelBuffer? {
+        let attrs = [kCVPixelBufferCGImageCompatibilityKey: kCFBooleanTrue!,
+                     kCVPixelBufferCGBitmapContextCompatibilityKey: kCFBooleanTrue!] as CFDictionary
+        var pixelBuffer: CVPixelBuffer?
+
+        let width = Int(self.size.width)
+        let height = Int(self.size.height)
+
+        let status = CVPixelBufferCreate(kCFAllocatorDefault,
+                                         width,
+                                         height,
+                                         kCVPixelFormatType_32ARGB,
+                                         attrs,
+                                         &pixelBuffer)
+
+        guard let buffer = pixelBuffer, status == kCVReturnSuccess else {
+            return nil
+        }
+
+        CVPixelBufferLockBaseAddress(buffer, [])
+
+        let colorSpace = CGColorSpaceCreateDeviceRGB()
+        let context = CGContext(data: CVPixelBufferGetBaseAddress(buffer),
+                                width: width,
+                                height: height,
+                                bitsPerComponent: 8,
+                                bytesPerRow: CVPixelBufferGetBytesPerRow(buffer),
+                                space: colorSpace,
+                                bitmapInfo: CGImageAlphaInfo.noneSkipFirst.rawValue)
+
+        guard let cgImage = self.cgImage else { return nil }
+
+        context?.draw(cgImage, in: CGRect(x: 0, y: 0, width: width, height: height))
+        CVPixelBufferUnlockBaseAddress(buffer, [])
+
+        return buffer
+    }
+}
